Done! Slim CI!

+++ dbt/models/base/base_meal_ingredients.sql
@@ -1,6 +1,7 @@
 -- ------------------------------------------------------------------------------
 -- Model: Base_meal_ingredients
--- Description: Base Table for meal ingredients
+-- Description: Base Table for meal ingredients - Only Needed for Ingredient ID (Int)
+-- Fast Comparison
 -- ------------------------------------------------------------------------------
 -- Change Log:
 -- Date       | Author   | Description
@@ -9,5 +10,5 @@
 -- YYYY-MM-DD | NAME     | [Add future changes here]
 -- ------------------------------------------------------------------------------
 
-SELECT id_ingredient as ingredient_id, str_ingredient as ingredient_name
+SELECT id_ingredient::INTEGER as ingredient_id, str_ingredient as ingredient_name
 FROM {{ source("meals", "ingredients") }}
